pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyupload"
version = "0.1.0"
description = "A minimal project description"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "tortoise-orm[asyncmy]>=0.25.3, <0.26.0",
    "uvicorn>=0.40.0, <0.41.0",
    "fastapi>=0.128.0, <0.129.0",
    "aerich>=0.9.2, <0.10.0",
    "dotenv>=0.9.9, <0.9.10",
    "bcrypt>=5.0.0, <5.1.0",
    "passlib>=1.7.4, <1.8.0",
    "starlette-sessions>=0.3.0, <0.4.0",
    "pydantic[email]>=2.12.5, <2.13.0",
    "itsdangerous>=2.2.0, <2.3.0",
    "pyjwt>=2.10.1, <2.11.0",
]

[project.scripts]
generate_users = "app.cli.generate_users:cli"

[project.optional-dependencies]
dev = [
    "fastapi[standard]>=0.128.0, <0.129.0",
]
test = [
    "pytest>=9.0.2, <10.0.0",
    "pytest-asyncio>=1.3.0, <2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.aerich]
tortoise_orm = "app.models.TORTOISE_ORM"
location = "./app/models/migrations"
src_folder = "."
